feat: reset marks user anonymous
